Keep the same type as the definiation in DDS spec

+++ rmw/include/rmw/events_statuses/matched.h
@@ -15,6 +15,7 @@
 #ifndef RMW__EVENTS_STATUSES__MATCHED_H_
 #define RMW__EVENTS_STATUSES__MATCHED_H_
 
+#include <stdint.h>
 #include <stddef.h>
 
 #include "rmw/visibility_control.h"
@@ -33,14 +34,14 @@ typedef struct RMW_PUBLIC_TYPE rmw_matched_status_s
    * For subscriber, total cumulative count the concerned subscriber discovered a 'match' with
    * a publisher since the last time the status was read.
    */
-  size_t total_count_change;
+  int32_t total_count_change;
 
   /**
    * For publisher, the number of subscribers currently matched to the concerned publisher.
    *
    * For subscriber, the number of publishers currently matched to the concerned subscriber.
    */
-  size_t current_count;
+  int32_t current_count;
 
   /**
    * The change in current_count since the last time the status was read.
@@ -48,7 +49,7 @@ typedef struct RMW_PUBLIC_TYPE rmw_matched_status_s
    * 'total_count_change - current_count_change' is the change for unmatched count since the last
    * time the status was read.
    */
-  size_t current_count_change;
+  int32_t current_count_change;
 } rmw_matched_status_t;
 
 #ifdef __cplusplus
